requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
email-validator==2.0.0
Werkzeug==2.2.3
SQLAlchemy==1.4.49
WTForms==3.0.1
python-dotenv==1.0.0
alembic==1.11.1 